requirements.txt
Pillow
numpy
imageio